1モデル1コントローラーへの変更に伴ってコントローラー作成
