--- app/Models/Address.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class Address extends Model
-{
-    use HasFactory;
-    protected $fillable = ['user_id','postal_code','prefecture','city','town','block','building'];
-}
