--- app/Models/Name.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class Name extends Model
-{
-    use HasFactory;
-    protected $fillable = ['user_id','name','kana_name', 'nickname'];
-}
