+++ app/Models/User.php
@@ -8,5 +8,20 @@ use Illuminate\Database\Eloquent\Model;
 class User extends Model
 {
     use HasFactory;
-    protected $fillable = ['name','email','gender','birthday','phone_number','img_path','kana_name','nickname'];
+    protected $fillable = [
+        'name',
+        'email',
+        'gender',
+        'birthday',
+        'phone_number',
+        'img_path',
+        'kana_name',
+        'nickname',
+        'postal_code',
+        'prefecture',
+        'city',
+        'town',
+        'block',
+        'building'
+    ];
 }
