+++ routes/web.php
@@ -1,8 +1,11 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\TaskController;
-use App\Http\Controllers\RegisterController;
+use App\Http\Controllers\{
+    TaskController,
+    RegisterController,
+    MailAuthenticationController
+};
 
 
 Route::get('/', function () {
@@ -11,8 +14,8 @@ Route::get('/', function () {
 Route::get('/tasks', [TaskController::class, 'index']);
 Route::post('/task',[TaskController::class,'store']);
 Route::delete('/task/{task}',[TaskController::class,'destroy']);
-Route::get('/register',[RegisterController::class,'index']);
-Route::post('/register',[RegisterController::class,'sendMail']);
+Route::get('/register',[MailAuthenticationController::class,'index']);
+Route::post('/register',[MailAuthenticationController::class,'sendMail']);
 Route::post('/create_user',[RegisterController::class,'register']);
 Route::get('/create_user/{token}',[RegisterController::class,'checkToken']);
